requirements.txt
Django==5.2.5
psycopg2-binary==2.9.10
Pillow==11.3.0
djangorestframework==3.16.1
django-filter==25.1
django-cors-headers==4.7.0
python-dotenv==1.1.1
dj-database-url==3.0.1
whitenoise==6.9.0
drf-spectacular==0.28.0
PyJWT==2.10.1
djangorestframework-simplejwt==5.5.1
requests==2.32.5
gunicorn==23.0.0
mysqlclient==2.2.5
